InchIToStructure and AuxInfoToStructure now both in same class;
